Rebase the UI-redesign branch + refactor old report activity and fragments
