+++ onebusaway-android/src/main/res/layout/report_issue_description_item.xml
@@ -25,7 +25,7 @@
             android:orientation="horizontal">
 
         <ImageView style="@style/MaterialSmallIcon"
-                   android:src="@drawable/ic_action_info"
+                   android:src="@drawable/ic_stop_info"
                    android:id="@+id/ic_action_info"/>
 
         <TextView
