+++ onebusaway-android/src/main/res/layout/report_stop_problem.xml
@@ -102,7 +102,7 @@
                 android:orientation="horizontal">
 
             <ImageView style="@style/MaterialSmallIcon"
-                       android:src="@drawable/ic_action_info"
+                       android:src="@drawable/ic_stop_info"
                        android:id="@+id/ic_action_info"/>
 
             <TextView
