+++ onebusaway-android/src/main/res/layout/report_trip_problem.xml
@@ -138,7 +138,7 @@
                 android:orientation="horizontal">
 
             <ImageView style="@style/MaterialSmallIcon"
-                       android:src="@drawable/ic_action_info"
+                       android:src="@drawable/ic_stop_info"
                        android:id="@+id/ic_trip_info"/>
 
             <TextView
