+++ onebusaway-android/src/main/res/layout/simple_arrival_list.xml
@@ -31,7 +31,7 @@
             android:orientation="horizontal">
 
         <ImageView style="@style/MaterialSmallIcon"
-                   android:src="@drawable/ic_action_info"
+                   android:src="@drawable/ic_stop_info"
                    android:id="@+id/arrival_list_action_info"/>
 
         <TextView
